pyproject.toml
[tool.poetry]
name = "unsloth"
version = "0.1.0"
description = "2-5X faster LLM finetuning"
authors = ["Unsloth AI team <info@unsloth.ai>"]
license = "MIT"
readme = "README.md"
homepage = "http://www.unsloth.ai"
repository = "https://github.com/unslothai/unsloth"
documentation = "https://github.com/unslothai/unsloth"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
setuptools = "*"
setuptools-scm = "*"
unsloth_zoo = "*"
cut-cross-entropy = "*"
packaging = "*"
tyro = "*"
transformers = "*"
datasets = "*"
sentencepiece = "*"
tqdm = "*"
psutil = "*"
wheel = "*"
numpy = "*"
accelerate = "*"
trl = "*"
peft = "*"
protobuf = "<4.0.0"
huggingface_hub = "*"
hf_transfer = "*"
bitsandbytes = "*"
xformers = "*"
torch = {version = "*", extras = ["cuda"]}
modelscope = "*"
python-dotenv = "*"
pytest = "*"
pytest-mock = "*"
# flash-attn = "*"  # Windows 環境下難以安裝，僅測試時可移除
triton-windows = "*"  # Windows 專用版本
torchvision = "*"
torchaudio = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
